Fix regression with music fade transitions not working
Unfortunately, 1de459d9b414180d5049cbf153266a8d6e00c227 caused a
regression where musicclass::niceplay() didn't work, because fading out
the music would cause haltdasmusik() to be called, which would reset the
fade variables.

To fix this, haltdasmusik() will now only reset the fade variables if
it's not called from a fade, which is signaled with a function
parameter.